Заменил тип поля Log.time с int на long

+++ week4/src/ru/croc/task9/Main.java
@@ -9,7 +9,7 @@ import java.util.Scanner;
 
 public class Main {
     static class Log {
-        int time;
+        long time;
         String msg;
         Log(String log) {
             String[] strings = log.split(" ");
